[fix] WeightedObject weight type

+++ project/wfc/wobj.py
@@ -3,4 +3,4 @@ from dataclasses import dataclass, field
 
 @dataclass(unsafe_hash=True)
 class WeightedObject:
-    weight: int = field(repr=False)
+    weight: float = field(repr=False)
